feat: place data json file 추가

See merge request s06-bigdata-rec-sub2/S06P22B203!8

+++ backend/src/main/java/com/b203/trou/entity/place/Place.java
@@ -25,8 +25,8 @@ public class Place extends BaseEntity {
     private String contentTypeId;
     private String firstImage;
     private String firstImage2;
-    private BigDecimal mapX;
-    private BigDecimal mapY;
+    private double mapX;
+    private double mapY;
     private int readCount;
     private String sigunguCode;
     private String tel;
